requirements.txt
accelerate==1.11.0
annotated-doc==0.0.3
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
attrs==25.4.0
beautifulsoup4==4.14.2
bitsandbytes==0.42.0
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorlog==6.10.1
cryptography==46.0.3
dill==0.4.0
distro==1.9.0
docling==2.60.0
docling-core==2.50.0
docling-ibm-models==3.10.2
docling-parse==4.7.0
docx2txt==0.9
et_xmlfile==2.0.0
faiss-cpu==1.12.0
Faker==37.12.0
fastapi==0.120.1
filelock==3.20.0
filetype==1.2.0
Flask==3.1.2
flask-cors==6.0.1
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
jsonlines==4.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
latex2mathml==3.78.1
lxml==6.0.2
markdown-it-py==4.0.0
marko==2.2.1
MarkupSafe==3.0.3
mdurl==0.1.2
mpire==2.10.2
mpmath==1.3.0
multiprocess==0.70.18
networkx==3.5
numpy==2.2.6
ocrmac==1.0.0
omegaconf==2.3.0
openai==2.6.1
opencv-python==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pathlib==1.0.1
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
pluggy==1.6.0
polyfactory==2.22.3
psutil==7.1.2
pyclipper==1.3.0.post6
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
Pygments==2.19.2
pylatexenc==2.10
PyMuPDF==1.26.5
pyobjc-core==12.0
pyobjc-framework-Cocoa==12.0
pyobjc-framework-CoreML==12.0
pyobjc-framework-Quartz==12.0
pyobjc-framework-Vision==12.0
pypdf==6.1.3
PyPDF2==3.0.1
pypdfium2==4.30.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.3
rapidocr==3.4.2
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
rich==14.2.0
rpds-py==0.28.0
rtree==1.4.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
semchunk==2.2.2
sentence-transformers==5.1.2
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
starlette==0.49.1
sympy==1.14.0
tabulate==0.9.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.9.0
torchvision==0.24.0
tqdm==4.67.1
transformers==4.57.1
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
Werkzeug==3.1.3
wheel==0.45.1
xlsxwriter==3.2.9
